fix(hotfix/v0.1.1): Wave 2 frontend — fix hasPairingFlow channel ID, add QR timeout
- ChannelConfigPanel.tsx: hasPairingFlow and isWhatsApp now check
  WHATSAPP_NATIVE_CHANNEL_ID ('whatsapp_native') so the panel stays
  open after Save & Enable and shows the QR notice for the native channel
- WHATSAPP_NATIVE_CHANNEL_ID constant (whatsappChannelId.ts) replaces
  three unbound 'whatsapp_native' string literals across two files
- WhatsAppPairingBody extracted to its own file (WhatsAppPairingBody.tsx)
- WhatsAppNativeNotice extracted to its own file (WhatsAppNativeNotice.tsx)
  with 15s initial QR timeout: if no QR frame arrives within 15s the UI
  surfaces a timeout+Retry state instead of an infinite spinner (#368)
- channel-fields.ts: getChannelFields normalises 'whatsapp_native' → 'whatsapp'
  so the panel receives the correct field descriptor for both channel ID forms
- ChannelConfigPanel.test.tsx: all renderPanel('whatsapp') calls updated to
  renderPanel('whatsapp_native') to match the new constant; 26/26 tests pass

Fixes #368 (partial — backend fix in parallel)
Fixes #283 (partial — backend fix in parallel)
