+++ src/components/skills/ChannelConfigPanel.test.tsx
@@ -175,7 +175,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC1: no frame → shows spinner with "Generating your QR code…"', async () => {
     // No pairing frame in store → waiting state (the default when the notice mounts).
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Generating your QR code/i)).toBeInTheDocument()
     })
@@ -183,7 +183,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC2: state "waiting" → shows spinner "Generating your QR code…"', async () => {
     mockPairingState('waiting')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Generating your QR code/i)).toBeInTheDocument()
     })
@@ -192,7 +192,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC2: state "code" → renders QR + Linked Devices steps + refresh note', async () => {
     mockPairingState('code', 'https://example.com/test-qr')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByTestId('whatsapp-qr')).toBeInTheDocument()
     })
@@ -202,7 +202,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC3: state "linked" → success message', async () => {
     mockPairingState('linked')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Linked successfully/i)).toBeInTheDocument()
     })
@@ -211,7 +211,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC4: state "timeout" → distinct expired copy + Retry button', async () => {
     mockPairingState('timeout')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/QR expired/i)).toBeInTheDocument()
     })
@@ -220,7 +220,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
 
   it('AC4: state "error" → distinct pairing-failed copy + Retry button', async () => {
     mockPairingState('error')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Pairing failed/i)).toBeInTheDocument()
     })
@@ -230,7 +230,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
   it('AC4: timeout and error render distinct copy from each other', async () => {
     // First render timeout
     mockPairingState('timeout')
-    const { unmount } = renderPanel('whatsapp', 'WhatsApp')
+    const { unmount } = renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/QR expired/i)).toBeInTheDocument()
     })
@@ -244,7 +244,7 @@ describe('ChannelConfigPanel — WhatsApp QR 5-state machine (#325 / US-C3)', ()
     vi.mocked(getChannelRouting).mockResolvedValue({ default_agent_id: undefined })
     vi.mocked(fetchAgents).mockResolvedValue([])
     mockPairingState('error')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Pairing failed/i)).toBeInTheDocument()
     })
@@ -317,7 +317,7 @@ describe('ChannelConfigPanel — WhatsApp Retry bounded timeout (MAJOR fix)', ()
         selector({ byChannel: pairingByChannel, clear: clearFn, apply: applyFn })) as never,
     )
 
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/QR expired/i)).toBeInTheDocument()
     })
@@ -353,7 +353,7 @@ describe('ChannelConfigPanel — WhatsApp Retry bounded timeout (MAJOR fix)', ()
         selector({ byChannel: pairingByChannel, clear: clearFn, apply: applyFn })) as never,
     )
 
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Pairing failed/i)).toBeInTheDocument()
     })
@@ -389,14 +389,14 @@ describe('ChannelConfigPanel — WhatsApp is always native (#283)', () => {
 
   it('renders the live linked-device pairing notice for whatsapp (no use_native gate)', async () => {
     // nativeAvailable is omitted → undefined → default-available.
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Generating your QR code/i)).toBeInTheDocument()
     })
   })
 
   it('does NOT render a use_native field (the toggle was removed)', async () => {
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByText(/Generating your QR code/i)).toBeInTheDocument()
     })
@@ -405,14 +405,14 @@ describe('ChannelConfigPanel — WhatsApp is always native (#283)', () => {
 
   it('renders the QR container when the pairing WS frame delivers a QR code', async () => {
     mockPairingState('code', 'https://example.com/test-qr')
-    renderPanel('whatsapp', 'WhatsApp')
+    renderPanel('whatsapp_native', 'WhatsApp')
     await waitFor(() => {
       expect(screen.getByTestId('whatsapp-qr')).toBeInTheDocument()
     })
   })
 
   it('renders the QR notice when native_available:true is explicitly passed', async () => {
-    renderPanel('whatsapp', 'WhatsApp', true)
+    renderPanel('whatsapp_native', 'WhatsApp', true)
     await waitFor(() => {
       expect(screen.getByText(/Generating your QR code/i)).toBeInTheDocument()
     })
@@ -421,7 +421,7 @@ describe('ChannelConfigPanel — WhatsApp is always native (#283)', () => {
 
   it('capability gating (#299): native_available:false shows the hint and NOT the QR notice', async () => {
     mockPairingState('code', 'https://example.com/test-qr')
-    renderPanel('whatsapp', 'WhatsApp', false)
+    renderPanel('whatsapp_native', 'WhatsApp', false)
     await waitFor(() => {
       expect(screen.getByTestId('native-unavailable-hint')).toBeInTheDocument()
     })
@@ -674,14 +674,14 @@ describe('ChannelConfigPanel — Save & Enable panel close behavior (#358)', () 
   })
 
   it('keeps the panel open after Save & Enable for WhatsApp (QR can render)', async () => {
-    const { onOpenChange } = renderPanel('whatsapp', 'WhatsApp', true)
+    const { onOpenChange } = renderPanel('whatsapp_native', 'WhatsApp', true)
 
     await act(async () => {
       fireEvent.click(screen.getByRole('button', { name: /Save & Enable/i }))
     })
 
     await waitFor(() => {
-      expect(vi.mocked(enableChannel)).toHaveBeenCalledWith('whatsapp')
+      expect(vi.mocked(enableChannel)).toHaveBeenCalledWith('whatsapp_native')
     })
     // The panel must NOT be closed — onOpenChange(false) would unmount the QR notice.
     expect(onOpenChange).not.toHaveBeenCalledWith(false)
