+++ src/components/skills/whatsappChannelId.ts
@@ -1,0 +1,2 @@
+/** Canonical channel_id for the native WhatsApp (whatsmeow) channel. */
+export const WHATSAPP_NATIVE_CHANNEL_ID = 'whatsapp_native'
